# backend/RefugeeShelter/src/main/resources/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "refugee-shelter-api",
    "description": "The API for the Refugee Shelter project",
    "contact": {
      "email": "akrikoff@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/AKRIKOFF/RefugeeShelter/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
    "/api/v1/rooms": {
      "get": {
        "tags": [
          "Room"
        ],
        "description": "Returns all rooms",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All rooms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/inline_response_200"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Room"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "in": "query",
            "description": "Current room owner",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1_rooms_body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Save room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/rooms/{roomId}": {
      "get": {
        "tags": [
          "Room"
        ],
        "description": "Return room by id",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "in": "path",
            "description": "Room ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_1"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Room"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "in": "path",
            "description": "Room ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/rooms_roomId_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Room"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "in": "path",
            "description": "Room ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_2"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Room"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "in": "path",
            "description": "Room ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/rooms_roomId_body_1"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_2"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/rooms/user/{ownerId}": {
      "get": {
        "tags": [
          "Room"
        ],
        "description": "Return all user's rooms",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "in": "path",
            "description": "Owner ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rooms by owner ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/inline_response_200_1"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/reservations/users/{userId}/rooms/{roomId}": {
      "post": {
        "tags": [
          "Reservation"
        ],
        "description": "Create reservation",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "roomId",
            "in": "path",
            "description": "Room ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/rooms_roomId_body_2"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservation by owner ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/inline_response_200_3"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "address": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "price": {
            "type": "integer"
          },
          "hasKitchen": {
            "type": "boolean"
          },
          "hasBathroom": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "inline_response_200": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "address": {
            "type": "string",
            "example": "Georgia, Tbilisi, Ivanova St, 8, 12"
          },
          "latitude": {
            "type": "number",
            "example": 32
          },
          "longitude": {
            "type": "number",
            "example": 123
          },
          "price": {
            "type": "integer",
            "example": 5300
          },
          "hasKitchen": {
            "type": "boolean",
            "example": true
          },
          "hasBathroom": {
            "type": "boolean",
            "example": false
          },
          "description": {
            "type": "string",
            "example": "This is really good flat for refugee from Ukraine And Russia"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "inline_response_403": {
        "type": "object",
        "properties": {
          "error_message": {
            "type": "string",
            "example": "The user is not logged in"
          }
        }
      },
      "v1_rooms_body": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "Georgia, Tbilisi, Ivanova St, 8, 12"
          },
          "latitude": {
            "type": "number",
            "example": 32
          },
          "longitude": {
            "type": "number",
            "example": 123
          },
          "price": {
            "type": "integer",
            "example": 5300
          },
          "hasKitchen": {
            "type": "boolean",
            "example": true
          },
          "hasBathroom": {
            "type": "boolean",
            "example": false
          },
          "description": {
            "type": "string",
            "example": "This is really good flat for refugee from Ukraine And Russia"
          }
        }
      },
      "inline_response_200_1": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "address": {
            "type": "string",
            "example": "Georgia, Tbilisi, Ivanova St, 8, 12"
          },
          "latitude": {
            "type": "number",
            "example": 32
          },
          "longitude": {
            "type": "number",
            "example": 123
          },
          "price": {
            "type": "integer",
            "example": 5300
          },
          "hasKitchen": {
            "type": "boolean",
            "example": true
          },
          "hasBathroom": {
            "type": "boolean",
            "example": false
          },
          "description": {
            "type": "string",
            "example": "This is really good flat for refugee from Ukraine And Russia"
          }
        }
      },
      "rooms_roomId_body": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "Georgia, Tbilisi, Ivanova St, 8, 12"
          },
          "latitude": {
            "type": "number",
            "example": 32
          },
          "longitude": {
            "type": "number",
            "example": 123
          },
          "price": {
            "type": "integer",
            "example": 5300
          },
          "hasKitchen": {
            "type": "boolean",
            "example": true
          },
          "hasBathroom": {
            "type": "boolean",
            "example": false
          },
          "description": {
            "type": "string",
            "example": "This is really good flat for refugee from Ukraine And Russia"
          }
        }
      },
      "inline_response_200_2": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "address": {
            "type": "string",
            "example": "Georgia, Tbilisi, Ivanova St, 8, 12"
          },
          "latitude": {
            "type": "number",
            "example": 32
          },
          "longitude": {
            "type": "number",
            "example": 123
          },
          "price": {
            "type": "integer",
            "example": 1000
          },
          "hasKitchen": {
            "type": "boolean",
            "example": true
          },
          "hasBathroom": {
            "type": "boolean",
            "example": false
          },
          "description": {
            "type": "string",
            "example": "This is really good flat for refugee from Ukraine And Russia"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "rooms_roomId_body_1": {
        "type": "object",
        "properties": {
          "price": {
            "type": "integer",
            "example": 1000
          }
        }
      },
      "inline_response_200_3": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 64
          },
          "startDate": {
            "type": "string",
            "example": "2022-01-01"
          },
          "endDate": {
            "type": "string",
            "example": "2022-01-01"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "rooms_roomId_body_2": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "example": "2022-01-01"
          },
          "endDate": {
            "type": "string",
            "example": "2022-01-01"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}